Worked more on Enemy Hitting, Its getting more and more consistant
Need to improve the Z amt when it is too low its not hit the intended spot and when it's too high It almost/barely goes out. Its something to do with the force multiplier and/or the need to add an extra z if necessary.

Also don't forget about bug where the enemy doesn't go to location because the hitting between the player and enemy is too quick for the predictballloc to find, idk if this is it cuz the predict func is very quick.

Also after finding a formula for the AI that works apply it to the player.