Minor fixes
Handles 4 cases:

Uses ICloud if the same, icloud ahead, or its a different user.

If its the same user and Icloud is behind then we use local save.

Also this is all building on the last commit. If all went right it should also be able to load cloud save even after deleting app.